backend/requirements.txt
flask
flask-cors
gunicorn
requests
sentence-transformers
faiss-cpu
scikit-learn
numpy
pandas
sqlalchemy
pickle5
